fixed resolve bug
